+++ compliant-reward-distribution/frontend/src/components/Admin/AdminGetAccountData.tsx
@@ -2,17 +2,18 @@ import { useState } from 'react';
 import { useForm, useWatch } from 'react-hook-form';
 import { Alert, Button, Form } from 'react-bootstrap';
 
-import { getARecentBlockHash, requestSignature, setClaimed, validateAccountAddress } from '../utils';
+import { getAccountData, getARecentBlockHash, requestSignature, validateAccountAddress } from '../../utils';
 import { ConcordiumGRPCClient } from '@concordium/web-sdk';
+import JSONbig from 'json-bigint';
 import { WalletProvider } from '../../wallet-connection';
 
 interface Props {
-    provider: WalletProvider | undefined;
     signer: string | undefined;
+    provider: WalletProvider | undefined;
     grpcClient: ConcordiumGRPCClient | undefined;
 }
 
-export function AdminSetClaimed(props: Props) {
+export function AdminGetAccountData(props: Props) {
     const { provider, signer, grpcClient } = props;
 
     interface FormType {
@@ -26,9 +27,11 @@ export function AdminSetClaimed(props: Props) {
     });
 
     const [error, setError] = useState<string | undefined>(undefined);
+    const [accountData, setAccountData] = useState<string | undefined>(undefined);
 
     async function onSubmit() {
         setError(undefined);
+        setAccountData(undefined);
 
         try {
             if (!signer) {
@@ -36,11 +39,11 @@ export function AdminSetClaimed(props: Props) {
             }
 
             const [recentBlockHash, recentBlockHeight] = await getARecentBlockHash(grpcClient);
+            const schema = 'FAADAAAADgAAAGNvbnRleHRfc3RyaW5nFgIHAAAAbWVzc2FnZQsKAAAAYmxvY2tfaGFzaBYC';
+            const signature = await requestSignature(recentBlockHash, schema, address, signer, provider);
 
-            const schema = 'FAADAAAADgAAAGNvbnRleHRfc3RyaW5nFgIHAAAAbWVzc2FnZRACCwoAAABibG9ja19oYXNoFgI=';
-            const signature = await requestSignature(recentBlockHash, schema, [address], signer, provider);
-
-            await setClaimed(signer, signature, recentBlockHeight, address);
+            const data = await getAccountData(signer, address, signature, recentBlockHeight);
+            setAccountData(JSONbig.stringify(data));
         } catch (error) {
             setError((error as Error).message);
         }
@@ -49,7 +52,7 @@ export function AdminSetClaimed(props: Props) {
     return (
         <div className="centered">
             <div className="card">
-                <h2 className="centered">Set Claimed</h2>
+                <h2 className="centered">Get Account Data</h2>
                 <br />
                 <Form onSubmit={handleSubmit(onSubmit)}>
                     <Form.Group className="col mb-3">
@@ -64,12 +67,18 @@ export function AdminSetClaimed(props: Props) {
                         <Form.Text />
                     </Form.Group>
                     <Button variant="secondary" type="submit">
-                        Set Claimed
+                        Get Account Data
                     </Button>
                 </Form>
-
+                <br />
                 {error && <Alert variant="danger">{error}</Alert>}
             </div>
+
+            {accountData && (
+                <div className="card">
+                    <pre className="pre">{JSON.stringify(JSON.parse(accountData), undefined, 2)}</pre>
+                </div>
+            )}
         </div>
     );
 }
